refactor(client): move IndexedDocument to separate module
- Move IndexedDocument struct and implementation from types.rs to
  new indexed_document.rs module
- Update imports in engine.rs and indexer.rs to use the new module
- Add proper documentation for the internal IndexedDocument type
- Keep IndexedDocument as internal-only type (not part of public API)

The change separates concerns by moving the internal intermediate
type to its own module, improving code organization.

feat(storage): re-export PageContent from persistence module

- Add PageContent to the re-exports in storage/mod.rs to make it
  available at the storage module level
