+++ rust/src/storage/mod.rs
@@ -42,5 +42,5 @@ mod persistence;
 pub mod workspace;
 
 // Re-export main types
-pub use persistence::{DocumentMeta, PersistedDocument};
+pub use persistence::{DocumentMeta, PageContent, PersistedDocument};
 pub use workspace::Workspace;
